Updated Aligned dataset tests
Added caltech birds script

+++ test/gan/test_aligned_dataset.py
@@ -19,7 +19,7 @@ from lernomatic.data import hdf5_dataset
 from lernomatic.util import file_util
 
 # debug
-from pudb import set_trace; set_trace()
+#from pudb import set_trace; set_trace()
 
 
 GLOBAL_OPTS = dict()
@@ -81,34 +81,104 @@ class TestAlignedDataSplit(unittest.TestCase):
 
 
 
-#class TestAlignedImageJoin(unittest.TestCase):
-#    def setUp(self):
-#        self.test_a_path = 'testA/'
-#        self.test_b_path = 'testB/'
-#        self.test_data_root = '/home/kreshnik/ml-data/monet2photo/'  # TODO: make settable
-#        # number of path pairs to place into split
-#        self.split_size  = 1024
-#        self.verbose     = GLOBAL_OPTS['verbose']
-#        # image properties
-#        self.test_image_shape = (3, 256, 256)
+class TestAlignedImageJoin(unittest.TestCase):
+    def setUp(self):
+        self.test_a_path = 'testA/'
+        self.test_b_path = 'testB/'
+        self.test_data_root = '/home/kreshnik/ml-data/monet2photo/'  # TODO: make settable
+        # number of path pairs to place into split
+        self.split_size  = 1024
+        self.verbose     = GLOBAL_OPTS['verbose']
+        # image properties
+        self.test_image_shape = (3, 256, 256)
+        self.test_out_image_shape = (3, 512, 256)
+
+    def test_proc(self):
+        print('======== TestAlignedImageJoin.test_proc ')
+
+        test_a_paths = file_util.get_file_paths(
+            str(self.test_data_root) + str(self.test_a_path),
+            verbose = self.verbose
+        )
+        self.assertNotEqual(0, len(test_a_paths))
+
+        test_b_paths = file_util.get_file_paths(
+            str(self.test_data_root) + str(self.test_b_path),
+            verbose = self.verbose
+        )
+        self.assertNotEqual(0, len(test_b_paths))
+        # For this test we need an equal amount of data
+        dataset_size = min(len(test_a_paths), len(test_b_paths))
+        test_a_paths = test_a_paths[0 : dataset_size]
+        test_b_paths = test_b_paths[0 : dataset_size]
+
+        print('Processing %d A images, %d B images from paths [%s]. [%s]' %\
+              (len(test_a_paths), len(test_b_paths),
+               str(self.test_data_root) + str(self.test_a_path),
+               str(self.test_data_root) + str(self.test_b_path))
+        )
+
+        test_image_dataset_name = 'images'
+        align_proc = aligned_data_proc.AlignedImageJoin(
+            image_dataset_name = test_image_dataset_name,
+            image_shape        = self.test_image_shape,
+            verbose            = self.verbose
+        )
+
+        test_outfile = 'data/test_aligned_image_join.h5'
+        align_proc.proc(test_a_paths, test_b_paths, test_outfile)
+
+        # check the dataset contents
+        with h5py.File(test_outfile, 'r') as fp:
+            dataset_keys = fp.keys()
+            self.assertIn(test_image_dataset_name, dataset_keys)
+            self.assertIn('a_ids', dataset_keys)
+            self.assertIn('b_ids', dataset_keys)
+
+        ## Test as HDF5Dataset
+        # Note that I am being sneaky here as I am loading the image as both
+        # the feature and the label. This is only done to shut the system up,
+        # since we don't actually need a label here. For the purpose of this
+        # test we care that the output is in fact twice the width (because that
+        # means that we joined the images correctly).
+        test_dataset = hdf5_dataset.HDF5Dataset(
+            test_outfile,
+            feature_name = align_proc.image_dataset_name,
+            label_name = align_proc.image_dataset_name,
+            verbose = self.verbose
+        )
+        self.assertEqual(len(test_a_paths), len(test_dataset))
+
+        # Check that the elements in the dataset are the expected size
+        for elem_idx, (feature, _) in enumerate(test_dataset):
+            print('Checking element [%d / %d]' % (elem_idx+1, len(test_dataset)), end='\r')
+            self.assertEqual(self.test_out_image_shape, feature.shape)
+
+        print('\n OK')
+
+        if GLOBAL_OPTS['remove']:
+            os.remove(test_outfile)
+
+        print('======== TestAlignedImageJoin.test_proc <END>')
 
 
 
 class TestAlignedImageSplit(unittest.TestCase):
     def setUp(self):
         self.test_data_root = "/mnt/ml-data/datasets/cyclegan/night2day/train/"
-        self.split_size     = 1024
+        self.test_size      = 4096
         self.verbose        = GLOBAL_OPTS['verbose']
         # image properties
         self.test_image_shape = (3, 256, 256)
 
-    def test_init(self):
-        print('======== TestAlignedImageSplit.test_init ')
+    def test_proc(self):
+        print('======== TestAlignedImageSplit.test_proc ')
 
         dataset_paths = file_util.get_file_paths(
             self.test_data_root,
             verbose = self.verbose
         )
+        dataset_paths = dataset_paths[0 : self.test_size]
         print('Processing %d images starting at root path [%s]' %\
               (len(dataset_paths), str(self.test_data_root))
         )
@@ -118,12 +188,11 @@ class TestAlignedImageSplit(unittest.TestCase):
         test_image_dataset_name = 'images'
         align_proc = aligned_data_proc.AlignedImageSplit(
             image_dataset_name = test_image_dataset_name,
-            image_shape = self.test_image_shape,
-            verbose = self.verbose
+            image_shape        = self.test_image_shape,
+            verbose            = self.verbose
         )
-        #self.assertEqual('AB', align_proc.direction)
 
-        test_outfile = 'data/test_aligned_proc.h5'
+        test_outfile = 'data/test_aligned_image_split.h5'
         align_proc.proc(dataset_paths, test_outfile)
 
         # Test as raw *.h5 file
@@ -151,9 +220,10 @@ class TestAlignedImageSplit(unittest.TestCase):
         print('\n OK')
 
         # now that the test is over, get rid of the file
-        os.remove(test_outfile)
+        if GLOBAL_OPTS['remove']:
+            os.remove(test_outfile)
 
-        print('======== TestAlignedImageSplit.test_init <END>')
+        print('======== TestAlignedImageSplit.test_proc <END>')
 
 
 
@@ -164,6 +234,11 @@ if __name__ == '__main__':
                         action='store_true',
                         default=False,
                         help='Sets verbose mode'
+                        )
+    parser.add_argument('--remove',
+                        action='store_true',
+                        default=False,
+                        help='Remove generated files at the end of tests'
                         )
     # data paths
     parser.add_argument('--test-data-root',
